+++ CHANGELOG.rst
@@ -12,10 +12,13 @@ This project adheres to `Semantic Versioning`_ starting with version 1.0.
 
 Added
 -----
+- ``PikaEventProducer`` adds the RabbitMQ ``App ID`` message property to published
+  messages with the value of the ``RASA_ENVIRONMENT`` environment variable. The
+  message property will not be assigned if this environment variable isn't set.
 
 Changed
 -------
-- updated mattermost connector documentation to be more clear.
+- Updated Mattermost connector documentation to be more clear.
 - Updated format strings to f-strings where appropriate.
 
 Removed